+++ src/main/java/org/example/Tables/OrderDishesDAO.java
@@ -10,7 +10,9 @@ import java.util.List;
 
 public class OrderDishesDAO {
 
-    public void addOrderDish(OrderDishes orderDish) {
+    public void addOrderDish(OrderDishes orderDish, String role) {
+        if (role.equals("User")) return;
+
         String sql = "INSERT INTO kp.order_dishes (id_order, dish_order, dish_amount_order) VALUES (?, ?, ?)";
 
         try (Connection conn = DataBaseConnection.getConnection();
@@ -24,7 +26,9 @@ public class OrderDishesDAO {
         }
     }
 
-    public void deleteOrderDish(int idOrder, int dishOrder) {
+    public void deleteOrderDish(int idOrder, int dishOrder, String role) {
+        if (role.equals("User")) return;
+
         String sql = "DELETE FROM kp.order_dishes WHERE id_order = ? AND dish_order = ?";
 
         try (Connection conn = DataBaseConnection.getConnection();
@@ -37,13 +41,17 @@ public class OrderDishesDAO {
         }
     }
 
-    public List<OrderDishes> getOrderDishes(int idOrder) {
+    public List<OrderDishes> getOrderDishes(int idOrder, String role) {
         List<OrderDishes> orderDishes = new ArrayList<>();
-        String sql = "SELECT * FROM kp.order_dishes WHERE id_order = ? AND dish_order = ?";
+        if (role.equals("User")) return orderDishes;
+
+        String sql = "SELECT * FROM kp.order_dishes WHERE id_order = ?;";
 
         try (Connection conn = DataBaseConnection.getConnection();
-             PreparedStatement stmt = conn.prepareStatement(sql);
-             ResultSet rs = stmt.executeQuery()) {
+             PreparedStatement stmt = conn.prepareStatement(sql);){
+            stmt.setInt(1, idOrder);
+            stmt.executeUpdate();
+            ResultSet rs = stmt.executeQuery();
             while (rs.next()) {
                 int dishOrder = rs.getInt("dish_order");
                 int dishAmountOrder = rs.getInt("dish_amount_order");
@@ -56,8 +64,10 @@ public class OrderDishesDAO {
         return orderDishes;
     }
 
-    public List<OrderDishes> getAllOrderDishes() {
+    public List<OrderDishes> getAllOrderDishes(String role) {
         List<OrderDishes> orderDishes = new ArrayList<>();
+        if (!role.equals("Admin")) return orderDishes;
+
         String sql = "SELECT * FROM kp.order_dishes";
 
         try (Connection conn = DataBaseConnection.getConnection();
