+++ src/main/java/org/example/Tables/UsersDAO.java
@@ -11,7 +11,8 @@ import java.util.List;
 
 public class UsersDAO {
 
-    public void deleteUser(int idUser) {
+    public void deleteUser(int idUser, String role) {
+        if (role.equals("User")) return;
         String sql = "DELETE FROM kp.users WHERE id_user = ?";
 
         try (Connection conn = DataBaseConnection.getConnection();
@@ -23,7 +24,8 @@ public class UsersDAO {
         }
     }
 
-    public void updateUserRole(int idUser, int updatedRole){
+    public void updateUserRole(int idUser, int updatedRole, String role){
+        if (role.equals("User")) return;
         if (updatedRole < 1 || updatedRole > 3){
             System.out.println("Ошибка!");
             return;
@@ -42,7 +44,8 @@ public class UsersDAO {
         }
     }
 
-    public void updateUserData(int idUser, String username, String passwordUser, String fullNameUser){
+    public void updateUserData(int idUser, String username, String passwordUser, String fullNameUser, String role){
+        if (role.equals("User")) return;
 
         String sql = "UPDATE kp.users\n" +
                 "SET username = "+username+"\n" +
@@ -60,7 +63,8 @@ public class UsersDAO {
         }
     }
 
-    public String getUser(int idUser){
+    public String getUser(int idUser, String role){
+        if (role.equals("User")) return "";
         String sql = "SELECT * FROM kp.users WHERE id_user = ?";
 
         int id = 0;
@@ -84,8 +88,12 @@ public class UsersDAO {
         return id + " " + username + " " + passwordUser + " " + fullNameUser + " " + roleUser;
     }
 
-    public List<String> getAllUsers() {
+    public List<String> getAllUsers(String role) {
+
         List<String> users = new ArrayList<>();
+
+        if (!role.equals("Admin")) return users;
+
         String sql = "SELECT * FROM kp.users";
         int idUser = 0;
         String username = "";
@@ -108,5 +116,24 @@ public class UsersDAO {
             e.printStackTrace();
         }
         return users;
+    }
+
+    public void updateSelfData(String username_now, String password_now, String username, String passwordUser, String fullNameUser){
+
+        String sql = "UPDATE kp.users\n" +
+                "SET username = "+username+"\n" +
+                "SET password_user = "+passwordUser+"\n" +
+                "SET full_name_user_user = "+fullNameUser+"\n" +
+                "WHERE username = ? AND password_user = ?";
+
+        try (Connection conn = DataBaseConnection.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, username_now);
+            stmt.setString(2, password_now);
+            stmt.executeUpdate();
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
     }
 }
